Smooth expected status chart
Rather than just taking the value at the beginning of each time period, now average all values in the time period. This won't change the result if the period is otherwise empty or if the beginning minute is the only one with a value, but now if there are multiple data points it will consider all of them.
This might not really be needed in practice, the chart isn't exactly high-precision. It might just be creating extra work for the computer that isn't visually apparent in the output.